api/requirements.txt
flask
requests
psycopg2-binary
boto3